requirements.txt
cryptography>=41.0.0
pyOpenSSL>=23.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
pytest>=7.0.0
requests>=2.31.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-jose>=3.3.0
aiohttp>=3.8.5
asyncio>=3.4.3
matplotlib>=3.7.0
pandas>=2.0.0
pycryptodome>=3.19.0
